pyproject.toml
[tool.poetry]
name = "orichalcum"
version = "0.1.0"
description = "YouTube data parser built for needlerank <https://github.com/nhtnr/needlerank>"
authors = ["nhtnr <noahtnr@gmail.com>"]
license = "MIT"

[tool.poetry.scripts]
orc = "orichalcum.orichalcum:main"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.27.1"
click = "^8.0.4"
python-dotenv = "^0.19.2"
psycopg = {extras = ["binary"], version = "^3.0.9"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
